+++ app/src/renderer/components/governance/LiProposal.vue
@@ -8,7 +8,7 @@ tr.li-proposal
         v-bind:class="status.color") checkmark
       span.validator-profile__status(v-else v-bind:class="status.color" v-tooltip.top="status.message")
       h2
-        router-link(:to="{ name: 'Proposal', params: { proposalId: proposal.proposal_id, proposal, status }}") {{ proposal.title }}
+        router-link(:to="{ name: 'Proposal', params: { proposalId: proposal.proposal_id }}") {{ proposal.title }}
       p.li-proposal__description {{ description }}
   td {{ `#` + proposal.proposal_id }}
   td.li-proposal__value.yes {{ proposal.tally_result.yes }}
@@ -30,30 +30,25 @@ export default {
     status() {
       if (this.proposal.proposal_status === `Passed`)
         return {
-          button: null,
           message: `This proposal has passed`
         }
       if (this.proposal.proposal_status === `Rejected`)
         return {
-          button: null,
           message: `This proposal has been rejected and voting is closed`,
           color: `red`
         }
       if (this.proposal.proposal_status === `DepositPeriod`)
         return {
-          button: `deposit`,
           message: `Deposits are open for this proposal`,
           color: `yellow`
         }
       if (this.proposal.proposal_status === `VotingPeriod`)
         return {
-          button: `vote`,
           message: `Voting for this proposal is open`,
           color: `green`
         }
       else
         return {
-          button: null,
           message: `There was an error determining the status of this proposal.`,
           color: `grey`
         }