+++ app/src/renderer/components/governance/TableProposals.vue
@@ -56,7 +56,6 @@ export default {
     somethingToSearch() {
       return !!this.proposals.length
     },
-    // TODO delete once tally is changed from Rat --> Dec
     parsedProposals() {
       if (!this.proposals || this.proposals.length === 0) return []